FIX(GDT): Fixed zeroed out GDT by manually copying data from LMA to VMA.
